Add files via upload
Childhood poverty has profound impacts on an individual's life and development in the future. Therefore, it is essential to analyze the geographical condition of childhood poverty for future policy advice. The data I choose in the first database is about children under age 5 living in poverty by county in New Jersey. The variables are counties in New Jersey, the estimated percentage of children under 5 who live in poverty divided by county in NJ, the numerator (number of children under 5 who live in poverty in a geographic area), and the denominator (the total number of children under 5 who live in a geographic location). The last two variables will not be used in ps1.
The education level is also significant to an individual's economic condition, and the education level of parents may influence children's living standards. My second chosen database is about the percentage of adults 25 and older with college completion rates 2016-2020 in NJ. The variables are the name of counties in NJ and the college completion rate in every geographic area.
My research question is "How does adults' education level influence the percent of children under 5 living in poverty in New Jersey from 2016-2020?" My hypothesis is "The counties with higher adult college completion rates have lower rates of children under 5 living in poverty in NJ from 2016-2020.